Trial 6.7

+++ src/estimator.py
@@ -23,13 +23,13 @@ def estimator(data):
     #CHALLENGE 2
 
     #For Impact
-    impact['severeCasesByRequestedTime'] = impact['infectionsByRequestedTime'] * 0.15
-    impact['hospitalBedsByRequestedTime'] = beds_available(totalHospitalBeds, impact['severeCasesByRequestedTime'])
+    impact['severeCasesByRequestedTime'] = int(impact['infectionsByRequestedTime'] * 0.15)
+    impact['hospitalBedsByRequestedTime'] = int(beds_available(totalHospitalBeds, impact['severeCasesByRequestedTime']))
 
 
     #For Severe Impact
-    severeImpact['severeCasesByRequestedTime'] = severeImpact['infectionsByRequestedTime'] * 0.15
-    severeImpact['hospitalBedsByRequestedTime'] = beds_available(totalHospitalBeds, severeImpact['severeCasesByRequestedTime'])
+    severeImpact['severeCasesByRequestedTime'] = int(severeImpact['infectionsByRequestedTime'] * 0.15)
+    severeImpact['hospitalBedsByRequestedTime'] = int(beds_available(totalHospitalBeds, severeImpact['severeCasesByRequestedTime']))
 
 
     output = {'data':data, 'impact':impact, 'severeImpact': severeImpact}
